try to have modules in eshop applications
