+++ eshop/core/pom.xml
@@ -9,15 +9,19 @@
         <version>0.0.1-SNAPSHOT</version>
     </parent>
 
-    <groupId>lt.codeacademy.security</groupId>
-    <artifactId>security-pom</artifactId>
-    <name>Security aggregator (pom)</name>
-    <packaging>pom</packaging>
+    <artifactId>core</artifactId>
+    <name>Core module (core)</name>
+    <version>0.0.1-SNAPSHOT</version>
 
-    <modules>
-        <module>security-core</module>
-        <module>disabled-security</module>
-        <module>basic-mvc</module>
-    </modules>
+    <dependencies>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-validation</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.projectlombok</groupId>
+            <artifactId>lombok</artifactId>
+        </dependency>
+    </dependencies>
 
 </project>
