--- eshop/main/src/main/java/lt/codeacademy/spring2025/eshop/common/validation/annotation/phonenumber/PhoneNumberPrefixType.java
@@ -1,14 +1,0 @@
-package lt.codeacademy.spring2025.eshop.common.validation.annotation.phonenumber;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Getter
-public enum PhoneNumberPrefixType {
-
-  GLOBAL("+370"),
-  LOCAL("0");
-
-  private final String prefix;
-}
