--- eshop/main/src/main/java/lt/codeacademy/spring2025/eshop/security/config/BasicSecurityConfig.java
@@ -1,81 +1,0 @@
-package lt.codeacademy.spring2025.eshop.security.config;
-
-import javax.sql.DataSource;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.log4j.Log4j2;
-import lt.codeacademy.spring2025.eshop.user.service.UserService;
-import org.springframework.boot.autoconfigure.security.servlet.PathRequest;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
-import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
-import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
-
-@Log4j2
-@Configuration
-@RequiredArgsConstructor
-@EnableMethodSecurity(prePostEnabled = true, securedEnabled = true, jsr250Enabled = true)
-@Profile("!no-security")
-public class BasicSecurityConfig {
-
-  private final DataSource datasource;
-  private final UserService userService;
-
-  @Bean
-  public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-    return http
-      .authorizeHttpRequests(authorizeRequests ->
-        authorizeRequests.requestMatchers(
-          "/",
-          "/login/**",
-          "/eshop_h2/**",
-          "/products/**", // changed from /products to /products/** for test security in controller level
-          "/products/find/**",
-          "/cart/**",
-          "/users/sign-up/**",
-          "/api/**" // allow access to API endpoints
-          ).permitAll()
-          .anyRequest()
-          .authenticated())
-      .formLogin(loginConfigurer ->
-        loginConfigurer.permitAll()
-          .loginPage("/login")
-          .loginProcessingUrl("/login")
-          .defaultSuccessUrl("/products", true)
-          .usernameParameter("loginEmail")
-          .passwordParameter("loginPassword")
-      )
-      .logout(logoutConfigurer -> logoutConfigurer
-        .logoutUrl("/logout")
-        .logoutSuccessUrl("/")
-        .logoutRequestMatcher(new AntPathRequestMatcher("/logout", "GET"))
-        .permitAll()
-      )
-      .build();
-  }
-
-  @Bean
-  public WebSecurityCustomizer webSecurityCustomizer() {
-    return web -> web.ignoring()
-      .requestMatchers(
-//        PathRequest.toH2Console(),
-        PathRequest.toStaticResources().atCommonLocations()  // add this line if static files prevented by security
-      );
-  }
-
-  @Bean
-  public AuthenticationProvider authenticationProvider() {
-    var daoAuthenticationProvider = new DaoAuthenticationProvider();
-    daoAuthenticationProvider.setUserDetailsService(userService);
-    daoAuthenticationProvider.setPasswordEncoder(new BCryptPasswordEncoder());
-
-    return daoAuthenticationProvider;
-  }
-}
