--- eshop/main/src/main/java/lt/codeacademy/spring2025/eshop/security/config/NoSecurityConfig.java
@@ -1,24 +1,0 @@
-package lt.codeacademy.spring2025.eshop.security.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.web.SecurityFilterChain;
-
-@Configuration
-@Profile("no-security")
-public class NoSecurityConfig {
-
-  @Bean
-  public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-    return http
-      .authorizeHttpRequests(authorizeRequests -> authorizeRequests
-        .anyRequest().permitAll() // Allow all requests without authentication
-      )
-      .csrf(AbstractHttpConfigurer::disable)
-      .cors(AbstractHttpConfigurer::disable)
-      .build();
-  }
-}
