--- eshop/main/src/main/java/lt/codeacademy/spring2025/eshop/security/domain/GlobalUserDto.java
@@ -1,16 +1,0 @@
-package lt.codeacademy.spring2025.eshop.security.domain;
-
-import java.util.StringJoiner;
-
-public record GlobalUserDto(String username,
-                            String password,
-                            String[] roles) {
-
-  @Override
-  public String toString() {
-    return new StringJoiner(", ", GlobalUserDto.class.getSimpleName() + "[", "]")
-      .add("username='" + username + "'")
-      .add("roles=" + String.join(", ", roles))
-      .toString();
-  }
-}
