--- eshop/main/src/main/java/lt/codeacademy/spring2025/eshop/user/dto/UserRoleDto.java
@@ -1,57 +1,0 @@
-package lt.codeacademy.spring2025.eshop.user.dto;
-
-import java.util.Collection;
-import java.util.HashSet;
-import java.util.Set;
-
-import lombok.Builder;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-@Builder
-public class UserRoleDto implements UserDetails {
-
-  private UserDto user;
-
-  @Builder.Default
-  private Set<? extends GrantedAuthority> roles = new HashSet<>();
-
-  @Override
-  public Collection<? extends GrantedAuthority> getAuthorities() {
-    return roles;
-  }
-
-  @Override
-  public String getPassword() {
-    return user.getPassword();
-  }
-
-  @Override
-  public String getUsername() {
-    return user.getEmail();
-  }
-
-  @Override
-  public boolean isAccountNonExpired() {
-    return UserDetails.super.isAccountNonExpired();
-  }
-
-  @Override
-  public boolean isAccountNonLocked() {
-    return UserDetails.super.isAccountNonLocked();
-  }
-
-  @Override
-  public boolean isCredentialsNonExpired() {
-    return UserDetails.super.isCredentialsNonExpired();
-  }
-
-  @Override
-  public boolean isEnabled() {
-    return UserDetails.super.isEnabled();
-  }
-
-  public String getFullName() {
-    return user.getFirstName() + " " + user.getLastName();
-  }
-}
