--- eshop/main/src/main/java/lt/codeacademy/spring2025/eshop/user/dto/UserSignUpDto.java
@@ -1,25 +1,0 @@
-package lt.codeacademy.spring2025.eshop.user.dto;
-
-import jakarta.validation.constraints.NotBlank;
-import lombok.Builder;
-import lt.codeacademy.spring2025.eshop.common.validation.annotation.fieldsstringcompare.FieldsStringCompare;
-import lt.codeacademy.spring2025.eshop.common.validation.annotation.phonenumber.PhoneNumber;
-import lt.codeacademy.spring2025.eshop.common.validation.annotation.phonenumber.PhoneNumberPrefixType;
-
-@Builder
-//@RepeatPassword
-@FieldsStringCompare(
-  firstField = "password",
-  secondField = "passwordRepeat",
-  message = "{user.signup.fields.not.match}")
-public record UserSignUpDto(@NotBlank String name,
-                            @NotBlank String surname,
-                            @NotBlank String email,
-                            @PhoneNumber(
-                              prefixType = PhoneNumberPrefixType.LOCAL,
-                              message = "Phone number length must be 9 characters"
-                            ) String phoneNumber,
-                            @NotBlank String zipCode,
-                            @NotBlank String city,
-                            @NotBlank String password,
-                            @NotBlank String passwordRepeat) {}
