--- eshop/main/src/main/java/lt/codeacademy/spring2025/eshop/user/model/AuthorityEntity.java
@@ -1,22 +1,0 @@
-package lt.codeacademy.spring2025.eshop.user.model;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity(name = "authority")
-@Table(name = "authority")
-public class AuthorityEntity {
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private long id;
-
-  @Column(nullable = false, unique = true)
-  private String name;
-  private String description;
-}
