+++ eshop/security/disabled-security/src/main/java/lt/codeacademy/security/config/NoSecurityConfig.java
@@ -1,0 +1,24 @@
+package lt.codeacademy.security.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.web.SecurityFilterChain;
+
+@Configuration
+@EnableMethodSecurity(prePostEnabled = false, securedEnabled = false, jsr250Enabled = false)
+public class NoSecurityConfig {
+
+  @Bean
+  public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+    return http
+      .authorizeHttpRequests(authorizeRequests -> authorizeRequests
+        .anyRequest().permitAll() // Allow all requests without authentication
+      )
+      .csrf(AbstractHttpConfigurer::disable)
+      .cors(AbstractHttpConfigurer::disable)
+      .build();
+  }
+}
