+++ eshop/security/security-core/src/main/java/lt/codeacademy/security/core/domain/AuthorityDomain.java
@@ -1,0 +1,18 @@
+package lt.codeacademy.spring2025.security.core.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+public class AuthorityDomain {
+
+  private Long id;
+  private String name;
+  private String description;
+
+}
