+++ eshop/security/security-core/src/main/java/lt/codeacademy/security/core/jpa/model/AuthorityEntity.java
@@ -1,0 +1,22 @@
+package lt.codeacademy.spring2025.security.core.jpa.model;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity(name = "authority")
+@Table(name = "authority")
+public class AuthorityEntity {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private long id;
+
+  @Column(nullable = false, unique = true)
+  private String name;
+  private String description;
+}
