+++ eshop/security/security-core/src/main/java/lt/codeacademy/security/core/mapper/AuthorityEntityDomainMapper.java
@@ -1,0 +1,28 @@
+package lt.codeacademy.spring2025.security.core.mapper;
+
+
+import org.springframework.stereotype.Component;
+
+import java.util.Set;
+import java.util.stream.Collectors;
+
+import lt.codeacademy.spring2025.security.core.domain.AuthorityDomain;
+import lt.codeacademy.spring2025.security.core.jpa.model.AuthorityEntity;
+
+@Component
+public class AuthorityEntityDomainMapper {
+
+  public AuthorityDomain mapToDomain(AuthorityEntity entity) {
+    return AuthorityDomain.builder()
+      .id(entity.getId())
+      .name(entity.getName())
+      .description(entity.getDescription())
+      .build();
+  }
+
+  public Set<AuthorityDomain> matToDomain(Set<AuthorityEntity> entitySet) {
+    return entitySet.stream()
+      .map(this::mapToDomain)
+      .collect(Collectors.toSet());
+  }
+}
